Add files via upload
This basic project attempts to learn/explore the best decision for a company to increase its sales revenue by deciding whether to improve customer experience on the website or app. 
The project applies linear regression to learn weights/coefficients of available features (lR^5) and on the basis of this suggests improvement of app experience to be a better decision.